requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
strawberry-graphql==0.96.0
strawberry-graphql-django==0.2.5
dj-database-url==0.5.0
gunicorn==20.1.0
python-decouple==3.6
psycopg2-binary==2.9.3
uvicorn==0.17.4
mypy==0.931
strawberry-django-plus==0.8.3